given the result.json filename, upload it, trajectories, and evaluation logs to google drive
