+++ sweagent/investigations/constants.py
@@ -1,0 +1,8 @@
+# We are focused on this one run (for now).
+DRIVE_REPRO_FOLDER_20240826 = "1ZkyW9o_KlHws3jpgBlgrDBrV31KOisYk"
+
+DRIVE_REPRO_FOLDER = "1TuQy5jJjocenNKZ_mlErB_Sx5xav_JCJ"
+
+DRIVE_DEFAULT_USER_INDEX = "2"
+
+RUN_LOGS_FOLDER_NAME = "run-logs"
